Produce real bundles on OSX.
1381e9e2 fixes Makefile to point to the right filename, which causes a
dynamic library to be used as a bundle. While the two are closely
related, they are not the same, although it does not cause problems in
practice (OSX originally could only load bundles at runtime, but since
10.4 it can handle both bundles and dylibs). The only practical
difference on a modern OSX system is probably file size - bundles are
considerably smaller.

This commit adds an extra linker pass that converts a dylib produced by
the rust compiler into a proper bundle.